Refactor main into modular handlers and core
